+++ ailets-rs/ailetos/tests/environment.rs
@@ -36,7 +36,7 @@ async fn test_value_node_writes_data_to_kv() {
         .expect("Failed to add value node");
 
     // Verify data was written to KV storage
-    let path = pipe_path(handle, actor_runtime::StdHandle::Stdout);
+    let path = pipe_path(handle, actor_runtime::StdHandle::Stdout as isize);
     let buffer = kv
         .open(&path, OpenMode::Read)
         .await
